94 - renaming test package
