+++ trade-module/src/test/java/io/edpn/backend/trade/adapter/persistence/stationlandingpadsizerequest/CreateStationLandingPadSizeRequestPortTest.java
@@ -1,8 +1,8 @@
-package io.edpn.backend.trade.adapter.persistence.stationrequireodysseyrequest;
+package io.edpn.backend.trade.adapter.persistence.stationlandingpadsizerequest;
 
 import io.edpn.backend.trade.adapter.persistence.StationLandingPadSizeRequestRepository;
 import io.edpn.backend.trade.adapter.persistence.repository.MybatisStationLandingPadSizeRequestRepository;
-import io.edpn.backend.trade.application.port.outgoing.stationlandingpadsizerequest.DeleteStationLandingPadSizeRequestPort;
+import io.edpn.backend.trade.application.port.outgoing.stationlandingpadsizerequest.CreateStationLandingPadSizeRequestPort;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -12,12 +12,12 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import static org.mockito.Mockito.verify;
 
 @ExtendWith(MockitoExtension.class)
-public class DeleteStationLandingPadSizeRequestPortTest {
+public class CreateStationLandingPadSizeRequestPortTest {
 
     @Mock
     private MybatisStationLandingPadSizeRequestRepository mybatisStationLandingPadSizeRequestRepository;
 
-    private DeleteStationLandingPadSizeRequestPort underTest;
+    private CreateStationLandingPadSizeRequestPort underTest;
 
     @BeforeEach
     public void setup() {
@@ -29,9 +29,9 @@ public class DeleteStationLandingPadSizeRequestPortTest {
         String systemName = "someName";
         String stationName = "someName2";
 
-        underTest.delete(systemName, stationName);
+        underTest.create(systemName, stationName);
 
-        verify(mybatisStationLandingPadSizeRequestRepository).delete(systemName, stationName);
+        verify(mybatisStationLandingPadSizeRequestRepository).insert(systemName, stationName);
     }
