+++ trade-module/src/test/java/io/edpn/backend/trade/adapter/persistence/stationlandingpadsizerequest/DeleteStationLandingPadSizeRequestPortTest.java
@@ -1,8 +1,8 @@
-package io.edpn.backend.trade.adapter.persistence.stationrequireodysseyrequest;
+package io.edpn.backend.trade.adapter.persistence.stationlandingpadsizerequest;
 
 import io.edpn.backend.trade.adapter.persistence.StationLandingPadSizeRequestRepository;
 import io.edpn.backend.trade.adapter.persistence.repository.MybatisStationLandingPadSizeRequestRepository;
-import io.edpn.backend.trade.application.port.outgoing.stationlandingpadsizerequest.CreateStationLandingPadSizeRequestPort;
+import io.edpn.backend.trade.application.port.outgoing.stationlandingpadsizerequest.DeleteStationLandingPadSizeRequestPort;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -12,12 +12,12 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import static org.mockito.Mockito.verify;
 
 @ExtendWith(MockitoExtension.class)
-public class CreateStationLandingPadSizeRequestPortTest {
+public class DeleteStationLandingPadSizeRequestPortTest {
 
     @Mock
     private MybatisStationLandingPadSizeRequestRepository mybatisStationLandingPadSizeRequestRepository;
 
-    private CreateStationLandingPadSizeRequestPort underTest;
+    private DeleteStationLandingPadSizeRequestPort underTest;
 
     @BeforeEach
     public void setup() {
@@ -29,9 +29,9 @@ public class CreateStationLandingPadSizeRequestPortTest {
         String systemName = "someName";
         String stationName = "someName2";
 
-        underTest.create(systemName, stationName);
+        underTest.delete(systemName, stationName);
 
-        verify(mybatisStationLandingPadSizeRequestRepository).insert(systemName, stationName);
+        verify(mybatisStationLandingPadSizeRequestRepository).delete(systemName, stationName);
     }
