+++ trade-module/src/test/java/io/edpn/backend/trade/adapter/persistence/stationlandingpadsizerequest/ExistsStationLandingPadSizeRequestPortTest.java
@@ -1,4 +1,4 @@
-package io.edpn.backend.trade.adapter.persistence.stationrequireodysseyrequest;
+package io.edpn.backend.trade.adapter.persistence.stationlandingpadsizerequest;
 
 import io.edpn.backend.trade.adapter.persistence.StationLandingPadSizeRequestRepository;
 import io.edpn.backend.trade.adapter.persistence.repository.MybatisStationLandingPadSizeRequestRepository;
